re-added missing 1.0-rapidia changes
Summary: Fixed missing content in 2.0.5r. This is the end result of rebasing 2.0.5r onto 5fce0ea.

Test Plan:
This should bring 2.0.5r more into alignment with the old 2.0.x branch, so the best way to test it is to
run through the usual test procedure document to verify that it does indeed behave holistically more like
the old 2.0.x branch.

Reading through the diff would also be appreciated.

Reviewers: tlai, edwinc

Reviewed By: edwinc

Differential Revision: https://phabricatorserver.rapidia.com/D95